+++ packages/sdk/vercel/src/createPlatformInfo.ts
@@ -1,7 +1,5 @@
 import type { Info, PlatformData, SdkData } from '@launchdarkly/js-server-sdk-common-edge';
 
-const packageJson = require('../package.json');
-
 class VercelPlatformInfo implements Info {
   platformData(): PlatformData {
     return {
@@ -11,8 +9,8 @@ class VercelPlatformInfo implements Info {
 
   sdkData(): SdkData {
     return {
-      name: packageJson.name,
-      version: packageJson.version,
+      name: '@launchdarkly/vercel-server-sdk',
+      version: '__LD_VERSION__',
     };
   }
 }