+++ packages/sdk/vercel/src/utils/mockEdgeConfigClient.ts
@@ -1,0 +1,10 @@
+import { EdgeConfigClient } from '@vercel/edge-config';
+
+const mockEdgeConfigClient: EdgeConfigClient = {
+  get: jest.fn(),
+  getAll: jest.fn(),
+  digest: jest.fn(),
+  has: jest.fn(),
+};
+
+export default mockEdgeConfigClient;