+++ packages/shared/akamai-edgeworker-sdk/src/platform/info/index.ts
@@ -1,0 +1,23 @@
+import type { Info, PlatformData, SdkData } from '@launchdarkly/js-server-sdk-common';
+
+class AkamaiPlatformInfo implements Info {
+  constructor(private platformName: string, private sdkName: string, private sdkVersion: string) {}
+
+  platformData(): PlatformData {
+    return {
+      name: this.platformName,
+    };
+  }
+
+  sdkData(): SdkData {
+    return {
+      name: this.sdkName,
+      version: this.sdkVersion,
+    };
+  }
+}
+
+const createPlatformInfo = (platformName: string, sdkName: string, sdkVersion: string) =>
+  new AkamaiPlatformInfo(platformName, sdkName, sdkVersion);
+
+export default createPlatformInfo;