+++ packages/sdk/vercel/CHANGELOG.md
@@ -2,6 +2,69 @@
 
 All notable changes to the LaunchDarkly SDK for Vercel Edge Config will be documented in this file. This project adheres to [Semantic Versioning](https://semver.org).
 
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @launchdarkly/js-server-sdk-common-edge bumped from 1.0.2 to 1.0.3
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @launchdarkly/js-server-sdk-common-edge bumped from 1.0.3 to 1.0.4
+
+## [0.4.1](https://github.com/launchdarkly/js-core/compare/vercel-server-sdk-v0.4.0...vercel-server-sdk-v0.4.1) (2023-06-07)
+
+
+### Bug Fixes
+
+* avoid modifying req.nextUrl ([#142](https://github.com/launchdarkly/js-core/issues/142)) ([49329d2](https://github.com/launchdarkly/js-core/commit/49329d2f142b83bc79361cd5b22c438f78a197b5))
+
+## [0.4.0](https://github.com/launchdarkly/js-core/compare/vercel-server-sdk-v0.3.0...vercel-server-sdk-v0.4.0) (2023-06-02)
+
+
+### Features
+
+* add new Vercel example ([#130](https://github.com/launchdarkly/js-core/issues/130)) ([d25f327](https://github.com/launchdarkly/js-core/commit/d25f327d9364ff3748e364426d9d6cfd83223bcf))
+* update Vercel example ([#129](https://github.com/launchdarkly/js-core/issues/129)) ([2296c4f](https://github.com/launchdarkly/js-core/commit/2296c4f8ad1febc3bd22c2272fcefebaf8d4cce6))
+
+## [0.3.0](https://github.com/launchdarkly/js-core/compare/vercel-server-sdk-v0.2.2...vercel-server-sdk-v0.3.0) (2023-05-16)
+
+
+### Features
+
+* Use version number from release process instead of package.json. ([#119](https://github.com/launchdarkly/js-core/issues/119)) ([9a5dcb8](https://github.com/launchdarkly/js-core/commit/9a5dcb8ef62756e278637c7e749cf2b204218d4a))
+
+## [0.2.2](https://github.com/launchdarkly/js-core/compare/vercel-server-sdk-v0.2.1...vercel-server-sdk-v0.2.2) (2023-04-28)
+
+
+### Bug Fixes
+
+* make vercel use EdgeProvider ([#114](https://github.com/launchdarkly/js-core/issues/114)) ([6705996](https://github.com/launchdarkly/js-core/commit/6705996929471ff8f72f97d58a665f75d4e5fecd))
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @launchdarkly/js-server-sdk-common-edge bumped from 1.0.1 to 1.0.2
+
+## [0.2.1](https://github.com/launchdarkly/js-core/compare/vercel-server-sdk-v0.2.0...vercel-server-sdk-v0.2.1) (2023-04-27)
+
+
+### Bug Fixes
+
+* add licence and fix missing package.json fields. ([c586398](https://github.com/launchdarkly/js-core/commit/c5863980c5bf4ee2a7590dfc4f7c575045d669b0))
+* remove beta text from cloudflare sdk readme. ([c586398](https://github.com/launchdarkly/js-core/commit/c5863980c5bf4ee2a7590dfc4f7c575045d669b0))
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @launchdarkly/js-server-sdk-common-edge bumped from 1.0.0 to 1.0.1
+
 ## [0.2.0](https://github.com/launchdarkly/js-core/compare/vercel-server-sdk-v0.1.1...vercel-server-sdk-v0.2.0) (2023-04-26)